fix: cannot read property 'removeFromDb' of undefined

+++ packages/core/database/src/collection.ts
@@ -1,7 +1,7 @@
 import merge from 'deepmerge';
 import { EventEmitter } from 'events';
 import { default as lodash, default as _ } from 'lodash';
-import { ModelCtor, ModelOptions, QueryInterfaceDropTableOptions, SyncOptions } from 'sequelize';
+import { ModelCtor, ModelOptions, QueryInterfaceDropTableOptions, SyncOptions, Transactionable } from 'sequelize';
 import { Database } from './database';
 import { Field, FieldOptions } from './fields';
 import { Model } from './model';
@@ -195,13 +195,19 @@ export class Collection<
   }
 
   async removeFromDb(options?: QueryInterfaceDropTableOptions) {
-    const queryInterface = this.db.sequelize.getQueryInterface();
-    await queryInterface.dropTable(this.model.tableName, options);
+    if (
+      await this.existsInDb({
+        transaction: options?.transaction,
+      })
+    ) {
+      const queryInterface = this.db.sequelize.getQueryInterface();
+      await queryInterface.dropTable(this.model.tableName, options);
+    }
     this.remove();
   }
 
-  async existsInDb() {
-    return this.db.collectionExistsInDb(this.name);
+  async existsInDb(options?: Transactionable) {
+    return this.db.collectionExistsInDb(this.name, options);
   }
 
   removeField(name) {
